backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
python-dotenv==1.0.1
openai==1.57.0
mysql-connector-python==9.1.0
sqlalchemy==2.0.36
python-multipart==0.0.18
